Refactored name of properties to match other models
