+++ ReactApp/src/models/Result.ts
@@ -1,7 +1,7 @@
 // Data class for result of a quiz as stored locally in MobX
 export class Result {
     constructor(
-        public quizId: string = "not set",
+        public kvisId: string = "not set",
         public answerResults: boolean[] = []
     ) {
     }
